requirement.txt

llvmlite>=0.38.0

numpy>=1.21.5
opencv-python>=4.5.5.64
Pillow>=9.1.0

tk>=0.1.0
tqdm>=4.64.0
